[Fix IntelliJ tests] Prevent class naming conflicts
Classes and package objects can’t share names between main, test and it.
